moved analyzer to sales analyst and tests
